app/src/main/java/com/example/beatbox/SoundViewModel.kt
package com.example.beatbox

import androidx.databinding.BaseObservable
import androidx.databinding.Bindable

class SoundViewModel(private val beatBox: BeatBox) : BaseObservable() {

    fun onButtonClicked() {
        sound?.let {
            beatBox.play(it)
        }
    }

    var sound: Sound? = null
        set(sound) {
            field = sound
            notifyChange()
        }

    @get:Bindable // yields BR.title in the BR class. short for Binding resources.
    val title: String?
        get() = sound?.name
}